startup refactor

+++ start.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
-mv ./frontend/env.example ./frontend/.env
-mv ./backend/env.example ./backend/.env 
-mv ./database/env.example ./database/.env 
+if [ -f ./frontend/env.example ]; then
+    mv ./frontend/env.example ./frontend/.env
+fi
+
+if [ -f ./backend/env.example ]; then
+    mv ./backend/env.example ./backend/.env 
+fi
+
+if [ -f ./database/env.example ]; then
+    mv ./database/env.example ./database/.env 
+fi
 
 sleep 2
 echo "Iniciando o projeto"
